Module Bin_prot

module Binable : sig ... end

module Blob : sig ... end

'a Blob.t is type-equivalent to 'a, but has different bin-prot serializers that prefix the representation with the size of 'a.

module Common : sig ... end

Common definitions used by binary protocol converters

module Md5 : sig ... end

module Nat0 : sig ... end

Nat0: natural numbers (including zero)

module Read : sig ... end

Reading values from the binary protocol using (mostly) OCaml.

module Shape : sig ... end

module Size : sig ... end

Compute size of values in the binary protocol.

module Std : sig ... end

This module defines default converters for the types defined in the OCaml standard library.

module Type_class : sig ... end

Sizers, writers, and readers in records

module Utils : sig ... end

module Write : sig ... end

Writing values to the binary protocol using (mostly) OCaml.