Module Utils.Make_binable_with_uuid

Parameters

module Bin_spec : Make_binable_with_uuid_spec

Signature

include Binable.S_only_functions with type t := Bin_spec.t

val bin_size_t : Bin_spec.t Size.sizer

val bin_write_t : Bin_spec.t Write.writer

val bin_read_t : Bin_spec.t Read.reader

val __bin_read_t__ : (int -> Bin_spec.t) Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Shape.t

val bin_writer_t : Bin_spec.t Type_class.writer

val bin_reader_t : Bin_spec.t Type_class.reader

val bin_t : Bin_spec.t Type_class.t